Fixed healthbar synchronization
Added a gradient option to change the color. Made calls for animations RPCs to  sync color flashes.
Fixed issue with health canvas slider not being updated.
Made canvas slider object not active initially and only active when loaded by a player to have the server not show it when loading.